# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minions Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with telegram code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/telegram:
    post:
      operationId: telegramAuth
      tags: [auth]
      summary: Authenticate via Telegram Mini App
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthBody"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/request-code:
    post:
      operationId: requestAuthCode
      tags: [auth]
      summary: Request auth code via Telegram bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCodeBody"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                required: [success]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        code:
          type: string
        telegramUsername:
          type: string
      required: [username, password, code]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    TelegramAuthBody:
      type: object
      properties:
        initData:
          type: string
      required: [initData]
    RequestCodeBody:
      type: object
      properties:
        telegramUsername:
          type: string
      required: [telegramUsername]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        bio:
          type: string
        balance:
          type: number
        frozenBalance:
          type: number
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        totalVolume:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        sellerLevel:
          type: string
        telegramId:
          type: string
        refCode:
          type: string
        createdAt:
          type: integer
        lastActive:
          type: integer
      required: [id, username]
...